final version

+++ contracts/BuildCollective.sol
@@ -210,7 +210,7 @@ contract BuildCollective is Ownable {
         bytes20 b = bytes20(keccak256(abi.encodePacked(msg.sender, now)));
         uint addr = 0;
         for (uint index = b.length-1; index+1 > 0; index--) {
-            addr += uint(b[index]) * ( 16 ** ((b.length - index - 1) * 2));
+            addr += uint(uint8(b[index])) * ( 16 ** ((b.length - index - 1) * 2));
         }
 
         return address(addr);
